+++ packages/marshal/test/marshal-justin.test.js
@@ -1,8 +1,9 @@
 import test from '@endo/ses-ava/prepare-endo.js';
 
-import { Remotable, makeTagged } from '@endo/pass-style';
+import { makeError, X } from '@endo/errors';
+import { Far, Remotable, makeTagged } from '@endo/pass-style';
 import { makeMarshal } from '../src/marshal.js';
-import { decodeToJustin } from '../src/marshal-justin.js';
+import { decodeToJustin, qp } from '../src/marshal-justin.js';
 import { jsonJustinPairs } from './_marshal-test-data.js';
 
 // this only includes the tests that do not use liveSlots
@@ -69,11 +70,39 @@ test('serialize decodeToJustin indented eval round trip', t => {
   });
   for (const [body, _, slots] of jsonJustinPairs) {
     const c = fakeJustinCompartment();
-    t.log(body);
     const encoding = JSON.parse(body);
     const justinExpr = decodeToJustin(encoding, true, slots);
+    t.log(justinExpr);
     const value = harden(c.evaluate(`(${justinExpr})`));
     const { body: newBody } = serialize(value);
     t.is(newBody, body);
   }
+});
+
+test('qp for quote passable as a quasi-quoted Justin expression', t => {
+  const r = Far('r', {});
+  const pr = Promise.resolve('fulfillment');
+  const e = makeError(X`xxx ${qp([3n, r, r, pr])} yyy ${qp(pr)}`);
+  t.log('zz', e, `ww`);
+  t.is(
+    e.message,
+    // In the literal string below, notice that the second call to `qp`
+    // starts the slot count over again. Otherwise, `qp` would be a
+    // communications channel. Since the rendered string is only for
+    // diagnostic value and does not show the slot content anyway,
+    // this is ok.
+    // TODO maybe better would be to show the slot contents using an outer
+    // `bestEffortsStringify` on the slots array. But that would be a
+    // terribly confusing mix of notation. We simply can't show them in
+    // Justin other than by slots.
+    `xxx ${JSON.stringify(`\`[
+  3n,
+  slotToVal("s0","Alleged: r"),
+  slotToVal("s0"),
+  slotToVal("s1"),
+]\``)} yyy "\`slotToVal(\\"s0\\")\`"`,
+  );
+
+  // For the original motivating example, which is also much more realistic,
+  // See qp-on-pattern.test.js in @endo/patterns
 });